# Faster Tree folds