Add some `Foldable` functions for `Data.Tree.Tree`.

* `foldMap`, `null`: Already exist
* `foldr`, `foldl'`: Added, basic functions
* `fold`, `elem`: Added, default definitions but `INLINABLE` helps
* `foldr1`, `foldl1`, `maximum`, `minimum`: Added, default definitions introduce `Maybe`s since these can be partial, we can avoid all that since trees are non-empty.
* `foldMap'`, `toList`, `length`, `sum`, `product`: Not added, defaults work well enough
* `foldr'`, `foldl`: Not added or benchmarked, unlikely to be used(?)

Benchmarks:

```
  fold:    OK (0.25s)
    15.5 μs ± 962 ns, 85% less than baseline
  foldMap: OK (0.14s)
    18.0 μs ± 1.7 μs, 82% less than baseline
  foldr:   OK (0.18s)
    21.8 μs ± 1.5 μs, 93% less than baseline
  foldl':  OK (0.30s)
    18.1 μs ± 878 ns, 95% less than baseline
  foldr1:  OK (0.16s)
    19.6 μs ± 1.5 μs, 94% less than baseline
  foldl1:  OK (0.30s)
    18.0 μs ± 1.2 μs, 95% less than baseline
  toList:  OK (0.17s)
    39.9 μs ± 2.9 μs, 31% less than baseline
  elem:    OK (0.27s)
    16.3 μs ± 777 ns, 47% less than baseline
  maximum: OK (0.16s)
    19.6 μs ± 1.5 μs, 94% less than baseline
  sum:     OK (0.30s)
    17.9 μs ± 671 ns, 95% less than baseline
```

Fixes #878.