@treeowl 2022-12-07
95% less than? So only 5% of the earlier time? That sounds ... A bit suspicious. Is it right?