@meooow25 2022-12-08
Unfortunately yes. Currently it seems to be ~20x slow for something like `sum`. I saw a similar difference when I reported the issue, and here's a comparison vs summing a list of the same size.
```
  sum Tree Int old: OK (0.21s)
    393  μs ±  23 μs
  sum [Int]:        OK (0.23s)
    13.9 μs ± 800 ns
```
```
  sum Tree Int new: OK (0.15s)
    18.2 μs ± 1.4 μs
```