@treeowl 2022-12-08
It's apparent that the trees in the test aren't big enough. I don't love comparing microseconds. Could you possibly bump those sizes up some and see what things look like?