@treeowl 2022-12-08
I don't know, really. Ideally we'd have a few sizes, but for now aim for half a second each, unless that makes the benchmarking take too long.