@meooow25 2022-12-08
Half a second would need something _very_ large. I've updated it to `2^20`, a little over a million.
```
  fold:    OK (0.76s)
    8.45 ms ± 366 μs, 76% less than baseline
  foldMap: OK (0.40s)
    8.89 ms ± 683 μs, 73% less than baseline
  foldr:   OK (0.42s)
    10.7 ms ± 753 μs, 94% less than baseline
  foldl':  OK (0.42s)
    10.5 ms ± 949 μs, 95% less than baseline
  foldr1:  OK (0.26s)
    10.0 ms ± 989 μs, 95% less than baseline
  foldl1:  OK (0.42s)
    10.5 ms ± 748 μs, 96% less than baseline
  toList:  OK (0.35s)
    30.0 ms ± 1.8 ms,  9% less than baseline
  elem:    OK (0.43s)
    10.3 ms ± 701 μs, 28% less than baseline
  maximum: OK (0.42s)
    10.5 ms ± 788 μs, 96% less than baseline
  sum:     OK (0.26s)
    10.5 ms ± 725 μs, 96% less than baseline
```