@meooow25 2022-12-21
Added benchmarks using a line tree as an example of a deep tree, since a binary tree is rather shallow. Also added a smaller size of 1000.

And I opened a GHC issue about the fold function: https://gitlab.haskell.org/ghc/ghc/-/issues/22659

<details>
<summary>New benchmarks</summary>
<pre>
  fold
    bin,n=1000:     OK (0.96s)
      4.20 μs ± 347 ns
    bin,n=1000000:  OK (0.58s)
      8.86 ms ± 820 μs
    line,n=1000:    OK (1.70s)
      5.30 μs ± 442 ns
    line,n=1000000: OK (0.56s)
      14.8 ms ± 802 μs
  foldMap
    bin,n=1000:     OK (1.74s)
      4.58 μs ± 338 ns
    bin,n=1000000:  OK (0.45s)
      8.93 ms ± 869 μs
    line,n=1000:    OK (1.77s)
      5.79 μs ± 367 ns
    line,n=1000000: OK (0.58s)
      15.2 ms ± 1.1 ms
  foldr_1
    bin,n=1000:     OK (1.75s)
      5.51 μs ± 331 ns
    bin,n=1000000:  OK (0.59s)
      10.7 ms ± 723 μs
    line,n=1000:    OK (1.58s)
      7.34 μs ± 667 ns
    line,n=1000000: OK (3.56s)
      46.1 ms ± 3.3 ms
  foldr_2
    bin,n=1000:     OK (1.59s)
      8.42 μs ± 716 ns
    bin,n=1000000:  OK (0.69s)
      17.0 ms ± 1.2 ms
    line,n=1000:    OK (1.42s)
      9.31 μs ± 757 ns
    line,n=1000000: OK (0.60s)
      34.6 ms ± 2.5 ms
  foldr_3
    bin,n=1000:     OK (1.63s)
      7.88 μs ± 674 ns
    bin,n=1000000:  OK (0.66s)
      15.0 ms ± 904 μs
    line,n=1000:    OK (1.78s)
      7.25 μs ± 365 ns
    line,n=1000000: OK (0.43s)
      28.5 ms ± 1.8 ms
  foldl'
    bin,n=1000:     OK (1.09s)
      4.43 μs ± 335 ns
    bin,n=1000000:  OK (0.61s)
      10.5 ms ± 753 μs
    line,n=1000:    OK (1.12s)
      5.81 μs ± 439 ns
    line,n=1000000: OK (0.78s)
      20.8 ms ± 1.1 ms
  foldr1
    bin,n=1000:     OK (2.08s)
      5.54 μs ± 207 ns
    bin,n=1000000:  OK (0.60s)
      10.1 ms ± 989 μs
    line,n=1000:    OK (1.78s)
      6.69 μs ± 443 ns
    line,n=1000000: OK (0.37s)
      19.6 ms ± 1.4 ms
  foldl1
    bin,n=1000:     OK (1.73s)
      4.35 μs ± 377 ns
    bin,n=1000000:  OK (0.87s)
      10.3 ms ± 419 μs
    line,n=1000:    OK (1.14s)
      6.27 μs ± 609 ns
    line,n=1000000: OK (2.51s)
      30.6 ms ± 1.6 ms
  toList
    bin,n=1000:     OK (1.61s)
      9.13 μs ± 761 ns
    bin,n=1000000:  OK (0.54s)
      29.4 ms ± 1.6 ms
    line,n=1000:    OK (1.61s)
      10.7 μs ± 814 ns
    line,n=1000000: OK (0.38s)
      49.4 ms ± 4.8 ms
  elem
    bin,n=1000:     OK (2.00s)
      4.48 μs ± 288 ns
    bin,n=1000000:  OK (0.88s)
      10.5 ms ± 440 μs
    line,n=1000:    OK (1.02s)
      5.50 μs ± 383 ns
    line,n=1000000: OK (0.37s)
      20.8 ms ± 1.7 ms
  maximum
    bin,n=1000:     OK (1.74s)
      4.52 μs ± 387 ns
    bin,n=1000000:  OK (0.37s)
      10.5 ms ± 974 μs
    line,n=1000:    OK (1.14s)
      6.16 μs ± 342 ns
    line,n=1000000: OK (0.79s)
      20.9 ms ± 722 μs
  sum
    bin,n=1000:     OK (1.99s)
      4.47 μs ± 285 ns
    bin,n=1000000:  OK (0.89s)
      10.3 ms ± 392 μs
    line,n=1000:    OK (1.15s)
      6.36 μs ± 388 ns
    line,n=1000000: OK (0.79s)
      21.0 ms ± 1.7 ms
</pre>
</details>