@treeowl 2022-12-28
Sorry, I got distracted with some lens stuff. I like folds, if they're clear and fast. I feel similarly about explicit recursion. Whatever's nicest in context.