@treeowl 2023-01-08
For `foldl1` and `foldr1`, I would recommend testing with the operation of a free magma.

```haskell
data Magma a
  = Inj a
  | Magma a :* Magma a
  deriving (Eq, Show)
```