@meooow25 2023-01-08
> Did you see my review?

Hi yes, was going through your comments.

> For `foldl1` and `foldr1`, I would recommend testing with the operation of a free magma.

Thanks, I was thinking the same but I wasn't sure what to name it. And in fact `foldr1` did give different results for a non-associative function, so thanks for catching this. Fixed now.